some more animations and refactoring
